requirements.txt
streamlit
groq
langchain
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
python-dotenv